Proof of concept: coroutine to implement iterator.
Implementing iterators is neither fun [1] nor easy. Even simple
operations, like the cross product, are much more difficult to implement
as an iterator than you would as a normal function. More complex
iterators like the merge part of a sort merge join are way, way more
complicated.

In my undergraduate compilers course, our professor mentioned that
coroutines are a good way to implement iterators. They not only make the
programs easier to write and read, but they are also typically very
fast. In fact, he argued that using them can often lead to code that's
faster than the hand-written code. I believe it; the hand written way is
kind of crazy.

In this commit, I use (one of) Boost's coroutine libraries [2] to
for a proof of concept to see if we can use coroutines to implement
iterators. I reimplemented the cross product using coroutines, and
things seem to work well! I think if I implemented a more complicated
iterator, the effects would be even more pronounced. Though things work,
I'm still not convinced coroutines are the way to go. The boost library
is a bit undocumented, and there are some performance penalties
associated with them [3]. So, we should think about it more.

[1]: Okay, it's pretty fun :P
[2]: http://www.boost.org/doc/libs/1_61_0/libs/coroutine2/doc/html/index.html
[3]: http://www.boost.org/doc/libs/1_61_0/libs/coroutine2/doc/html/coroutine2/performance.html