Introduced variadic templates into flow.
    "He that justifieth the wicked, and he that condemneth the just, even
    they both [are] abomination to the LORD."

        - Proverbs 17:15

I'm going to write a longer, clearer explanation about what variadic
templates are and why we're using them. For now, it suffices to say that
we're using this super complicated C++ feature to achieve a level of
type safety that I think is hard to achieve in pretty much any other
language. The code itself is an abomination unto the lord and limits the
amount of stuff we can do at runtime, but it avoids us having to write a
lot of complicated code while simultaneously catching a lot of bugs at
compile time. Is it worth it? Still up for discussion. This commit just
starts to show that it can be done.